move Storybook dir
