+++ package.json
@@ -11,9 +11,6 @@
   "sideEffects": [
     "lib/setLogger.js"
   ],
-  "engines": {
-    "npm": "use yarn instead of npm."
-  },
   "files": [
     "dist",
     "lib",
@@ -39,13 +36,8 @@
     "@babel/preset-env": "^7.9.0",
     "@babel/preset-typescript": "^7.10.4",
     "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^6.0.0",
+    "@rollup/plugin-node-resolve": "^11.2.1",
     "@rollup/plugin-typescript": "^6.0.0",
-    "@storybook/addon-actions": "^6.0.26",
-    "@storybook/addon-essentials": "^6.0.26",
-    "@storybook/addon-links": "^6.0.26",
-    "@storybook/addons": "^5.3.18",
-    "@storybook/svelte": "^6.0.26",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/svelte": "^3.0.0",
     "@testing-library/user-event": "^12.1.7",
@@ -53,7 +45,7 @@
     "@types/jest": "^26.0.15",
     "@typescript-eslint/eslint-plugin": "^3.6.1",
     "@typescript-eslint/parser": "^3.6.1",
-    "axios": "^0.20.0",
+    "axios": "^0.21.1",
     "babel-jest": "^25.2.6",
     "babel-loader": "^8.1.0",
     "cross-env": "^7.0.2",
@@ -72,7 +64,7 @@
     "graphql": "^15.3.0",
     "graphql-request": "^3.2.0",
     "husky": "^4.3.0",
-    "jest": "^25.2.7",
+    "jest": "^26.6.3",
     "nan": "^2.14.1",
     "path": "^0.12.7",
     "prettier": "^2.0.5",
@@ -85,9 +77,7 @@
     "svelte": "^3.29.0",
     "svelte-check": "^1.0.55",
     "svelte-jester": "^1.0.5",
-    "svelte-loader": "^2.13.6",
-    "svelte-preprocess": "^4.5.1",
-    "ts-jest": "^26.4.1",
+    "ts-jest": "^26.5.4",
     "type-fest": "^0.21.1",
     "typescript": "^4.0.3"
   },
@@ -96,7 +86,7 @@
     "react-query"
   ],
   "dependencies": {
-    "broadcast-channel": "^3.4.1"
+    "broadcast-channel": "^3.5.3"
   },
   "husky": {
     "hooks": {
