+++ storybook/package.json
@@ -1,0 +1,18 @@
+{
+    "name": "@sveltestack/svelte-query-storybook",
+    "version": "0.0.1",
+    "license": "MIT",
+    "scripts": {
+        "storybook": "start-storybook -p 6006",
+        "build-storybook": "build-storybook"
+    },
+    "devDependencies": {
+        "@storybook/addon-actions": "^6.2.4",
+        "@storybook/addon-essentials": "^6.2.4",
+        "@storybook/addon-links": "^6.2.4",
+        "@storybook/addons": "^6.2.4",
+        "@storybook/svelte": "^6.2.4",
+        "svelte-loader": "^3.1.0",
+        "svelte-preprocess": "^4.7.0"
+    }
+}
