+++ storybook/stories/query/App.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { QueryClientProvider } from '../../src'
+  import { QueryClientProvider } from '../../../src'
   import Query from './Query.svelte'
 </script>
