+++ test/queries.spec.ts
@@ -3,7 +3,7 @@
 import "@testing-library/jest-dom/extend-expect";
 import { render, waitFor } from "@testing-library/svelte";
 
-import Queries from "../stories/queries/App.svelte";
+import Queries from "../storybook/stories/queries/App.svelte";
 
 describe("Queries", () => {
   it("should fetch 2 queries", async () => {
